Gui Fully Implemented
Added a GuiManager that handles all of the necessary exposure in order to get imGUI to work with Vulkan.  It creates its own Vulkan resources, and is in charge of cleaning them up upon resize/termination similar to the swapchain.  There is a #defined bool in VulkanManager that you can change to false if you want to disable GUI.
There's some weirdness since the swapchain needs to know if you're going to initialize the GUI before it starts creating itself, and the GUI needs the swapChain to already be created.  So there's an extra bool in VulkanManager to account for this in lieu of passing in a boolean to SwapChain creation (which might be better now that I think about it)