Octree now functional with changing collider colors upond collision
NOthing in here is overly efficient, and I'll be editing for efficiency in the near future.
Given the physics system will be overhauled soon anyway, I figured just a proof of concept implementation was sufficient.
Known issues: When skybox is created, its collider is always at the center and clearly visible (since it's created as a physics object).  There should be a way to fully disable it as a physics object since we don't want to ever collide with it.
Known Issue 2: Framerate.  It tanks considerably while drawing handles with more than 100 objects.  Same goes for the Octree even with optimal spatial partitioning.  It can safely hover at aroun 20 FPS with 200 objects tho, and this is in Debug mode since Release is non functioning ATM.
TODO: Once physics and master are properly merged, rework Octree so that it's reconstructilbe at will (likely will do this anyway before then).