About to try some wacky stuff.  Might break everything so I'm makin gthis a restore point!
Attempting to make color instanced data.  I'll be changing the mesh AddInstance method to take in a color as well as a position and storing that into a struct.  From there, the instance buffer will be recreated using the size of the struct, not just Transform data.
Shoud color not be passed in, it will simply just use the mesh's default color (or white...?)