Added instanced Coloring
Kinda just works as a tint for the basic shader.  Currently considering how to apply it more generally without discrading original vertex color (i.e. the instanced color will overwrite the color stored on the vertex if it is the only color used.  Hence why it works bes as a tint.)